requirements.txt
Pillow==9.1.0
sentence_transformers==2.2.0
torch==1.11.0
tqdm==4.62.3
typing_extensions==4.1.1
